+++ Ui/Component/Listing/Columns/Status.php
@@ -1,8 +1,10 @@
 <?php
+
 /**
  * Copyright © Improntus All rights reserved.
  * See COPYING.txt for license details.
  */
+
 namespace Improntus\ProntoPaga\Ui\Component\Listing\Columns;
 
 use Magento\Framework\Data\OptionSourceInterface;
@@ -30,7 +32,8 @@ class Status implements OptionSourceInterface
                 ['value' => 'success', 'label' => __('Success')],
                 ['value' => 'canceled', 'label' => __('Canceled')],
                 ['value' => 'rejected', 'label' => __('Rejected')],
-                ['value' => 'pending', 'label' => __('Pending')],
+                ['value' => 'refunded', 'label' => __('refunded')],
+                ['value' => 'refund', 'label' => __('')],
                 ['value' => 'expired', 'label' => __('Expired')],
                 ['value' => 'error', 'label' => __('Error')]
             ];