+++ CHANGELOG.md
@@ -6,3 +6,13 @@ CHANGELOG
 ### 1.0.1 (2024-04-12)
 * Updated API request methods and remove 'message'.
 * Updated JS files and apply improvements by eslint.
+* Update UI listing to show request body an response payload parsed by html entity.
+
+### 1.0.2 (2024-04-14)
+* Two observer are created for when the credit memo of the order is generated and the refund is carried out through ProntoPaga.
+* Added configuration for the observer on system.xml
+* Update validate action from Pronto Paga transaction UI grid.
+* Added two new constants on Helper.
+* Added new values on config.xml.
+* Added new status `refuned`.
+* Fix css.